project/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["ameeraazli <wanazliwanameera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.7.0"
pandas = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
